move canvas fallback to separate file
